Fix compatibility with PHP 8.3.0
The imagerotate() function used to take 3 parameters. Then, in PHP
5.1.0, an optional fourth parameter was added. And now, in PHP 8.3.0,
that parameter has been removed. Since the code in GdAdapter always
explicitly provides the fourth parameter when calling the function,
this causes it to fail with an ArgumentCountError under PHP 8.3.0.

This commit removes the unused fourth parameter from all call sites.